frontend: type address signing
This is the last component that used the subscribe decorator.

React class componentDidUpdate is not necessary anymore, this was
called when the props changed and provided prevProps. Changed to
state / subscription, which means we don't have access to prevProps
which is rewritten using currentState and setState callback. This
part of the code is responsible for managing the account to use.