backend/bitbox02bootloader: introduce intermediate upgrades
**Add support for intermediate upgrades**

When going from version X to version Z, sometimes it is necessary to
first upgrade to an intermedaite version Y and run it. This commit
adds support for that, guiding the user through all required
intermediate upgrades before upgrading to the final latest version.

The user does not need to click through all intermediate upgrades -
they confirm the upgrade to the newest version, and the intermediate
upgrades are automatically flashed and booted in order.

**Add monotonic versions to the bundled firmware info**

They can be read from the binary (and the unit tests make sure they
match), but the code that needs this info is simpler if it is simply
available instead of needing to parse it out of the binary.