frontend: deactivate hide amounts if disabled from settings page
previously, we deactivate hide amounts only via the "hide/show
amounts" button (shown on the top of some pages).

Due to this, amounts will still be hidden even if we disable
the settings as a whole - given we didn't disable the feature
via the "hide/show amounts" btn.

This commit fixes it by always setting `setHideAmounts(false)`
as well as the config value `hideAmounts: false` if we disable
the feature via settings.