frontend: change to sort accounts in app component
Only sort accounts where they are received and pass to child
components. Before it was sorting the accounts on every sidebar
render. Now they are only sorted when app component updates.

It is currently not possible to sort directly only after useLoad
or useSubscribe, but it should be good enough. Sorting could
probabably be reduced with some clever useEffects if needed.