backend/keystore: detect new keystores after a wrong one
With the current implementation, when the wrong keystore is connected after
a keystore request, the app is not able to automatically detect the
connection of the correct keystore. The only option for the user is to
cancel the 'connect keystore' popup and repeat the action to create a new
keystore connection request. This is counterintuitive and inefficient.

This adds a `connectkeystore.retryCallback` that allows to wait for a new
keystore to be connected and to check if that is the expected one.