frontend: style failed transactions
ETH transactions can sometimes fail, in that case the entry in the
transaction list should be styled differently to clearly distinguish
it from other completed transactions.

This commit changes the transaction icon to warning symbol and
removes the color on the failed amount to be just gray instead of
green or red.