deps: switch back to the karalabe/hid USB library
When building the BBApp on an updated XCode, the BitBox02 can't
connect anymore on macOS, most likely due to
https://github.com/libusb/hidapi/issues/531.

In
https://github.com/digitalbitbox/bitbox-wallet-app/commit/bef0672b50ee29b0e92ac7f06087c233a58f5231
we switched from karalabe/hid to karalabe/usb, because karalabe/usb
go-ethereum switched to it, and both libs cannot (or could not) be
linked at once.

Then in
https://github.com/digitalbitbox/bitbox-wallet-app/commit/63dbb28007e1bb6e0bd67c456d20fc3f1a56e1ce
we abandoned karalabe/usb and made our own
fork (https://github.com/digitalbitbox/usb/), because it seemed
unmaintained/dead, and we needed important fixes to keep working.  The
original and the fork had a very old version of
https://github.com/libusb/hidapi/.

This fork, like the original, bundled a very old version of the hidapi
C library.

The above issue is with the connection on macOS is fixed by uprading
it, but that is not super easy (hidapi requires libusb, and upgrading
and compiling both in cgo is hard).

Luckily just last week, go-ethereum switched back to karalabe/hid, and
karalabe/hid finally got some love from the maintainers and they
upgraded hidapi:
https://github.com/karalabe/hid/commit/80d2c4b117fd6f12db2371f14b45cf287665078c

Switching back to this the hid lib fixes the macOS issue and it means
we can archive our own fork.

The golangci linter `gomoddirectives` is disabled b/c it complains
about `replace` directives in the go.mod file.