frontend/android: fix uploading KYC docs for moonpay
The link to upload KYC documents on moonpay wasn't working for Android.
This fixes the issue implementing an handler for the file picker. This
needs a new androidx.activity dependency to work reliably on different
Android versions, and also requires to upgrade androidx.appcompat to
1.6.1 to avoid build errors.

co-authored with <marko@shiftcrypto.ch>