frontend: reinitialize chart when chart data length changed
Sometimes the chart was only half filled or not everything was
displayed. This can happen if there is at least one rememebered
wallet in combination with an unlocked wallet with different age,
so that the length of the chart data changes.

calling fitContent after setting the new data fixes this
https://tradingview.github.io/lightweight-charts/docs/api/interfaces/ITimeScaleApi#fitcontent