frontend: use native locale for amount formatting
Before this change the amount formatting relied on the current
i18n.language to format values, but i18n.language changes when the
user selects a different language.

It should always format amoutns in the native locale, regardless
of the current langauge.

Changed to format by native locale and added nativeLocale to app
context as this does not change while the application is running.