frontend: convert manage-backups switch to functional component
This is a container for the manage-backup route and a switch to
render backup component for BB01/BB02. This is not used for setup
restore-from-backup, but only on the route when the user accesses
manage backups from settings.

The class methods were mostly used to render different parts of
the component. Moved this to separate components in the same file.

Also changed to named export instead of default export.