frontend: expose useLocalizedPunctuation in Context API
useLocalizedPunctuation may get noisy (log/warnings) if native
locale is empty or invalid. This is due to the hook
being used in many places and thus it runs indpendently
of one another.

Using context API, we can expose this globally and every
consumer will get the same instance of useLocalizedPunctuation.
This reduces the calls of useLocalizedPunctuation which results
in fixing the warning noises.