backend: update coin control outputs sorting
Currently, the list of all UTXOs in the coin control
selection is ordered by amount. Users sometimes re-use
addresses and are not aware that there might be other
UTXOs they didn't want or want to use on the same address
as their selected output. Grouping the UTXOs by address
helps the user to detect if there are other UTXOs.

This commit updates the sorting of outputs in the coin
control dialog and is now grouped by address from the
biggest to the lowest amount.