frontend: create dmg installer in osx build
Installing application with .dmg files is a familiar and intuitive
process for macOS users, ensuring applications are installed and
updated correctly.

Using create-dmg we can create a nice looking .dmg installer with
custom icons and background. This new dependancy is added to the
osx-brew.sh script and will also tried to be installed when running
osx-create-dmg.sh.

The Makefile is adjusted with a new osx-create-dmg target, which should
be run manually after the BitBox.app has been built and notarized.

Instructions to create the DMG installer have been added to the build
documentation in BUILD.md.