backend/account: migrate legacy notes upfront
In the past we had different account IDs, and the notes where stored
in different files. To deal with this, we used a list of notes objects
per account to be able to fetch notes from all files.

This commit changes this to migrate/merge legacy note files so we only
have one notes file per account. This is conceptually easier and
simplifies the upcoming notes export feature.

This also fixes a bug where the rates updater event is not propagated
for accounts with account number > 0 - Initialize() returned early
with `return nil` in that case.

NotesData is renamed to Data as the linter thougth it was a bad name:

> backend/accounts/notes/notes.go:30:6: exported: type name will be used as notes.NotesData by other packages, and that stutters; consider calling this Data (revive)