backend: add labels export & import
Export & export account names and transaction notes. Export & import
only touches the accounts of connected/remembered keystores.

We use BIP-329 so users can export the notes and import them into
other third party wallets that support BIP-329.

We add an extension to BIP-329 by adding a "bitboxapp" field - see
comments in commit for rationale.

We also don't use the "origin" field for now, as it is Bitcoin
specific, but we support LTC/ETH as well. We could add support for it
in the future anyway for the BTC-only entries.

If the bitboxapp field is present, the import is unambiguous - we know
clearly which account to target.

If the bitboxapp field is absent (i.e. import from a third party
wallet), it will generally work except for a few edge cases:

- If someone has different labels for multiple sub-accounts in a
unified account, the BitBoxApp will receive the name of one of
them (the last one in the file).
- If someone has different labels for the same transaction but in
different accounts, only one of them will receive label.

The import is currently somewhat inefficient, as it writes the whole
notes file of an account for each imported note by using
`SetTxNote()`. We can optimize this in a future PR if needed.

Conflict resolution: for now we simply prioritize the note in the
import over an existing note if there is a conflict. Users are not
expected to run into this a lot, as most use only one host device
anyway. We could add manual conflict resolution in the future if
needed.