frontend: improve vertical position of badges
Badges are not vertically aligned consistently in all places. This
was caused mostly as badges were centerd vertically but then had
slightly different line-height.

Badges behave like text and usually have `display: inline;`.

Changed line-height to be consistent and vertical alignment to
baseline, so that the badge text stands on the same line as the
other text.

Badges currently appear in the following components:
- account selector can have badge with icon only
- sidebar shows badge for connected keystore with icon and text
- reused utxo's in coincontrol have badge with warning text
- account summary balance have badge for connected keystore
- exchange options show best-deal or fast badge
- manage-accounts shows badge for connected keystore