frontend: fix sidebar and app-content not scrollable css regression
With iOS prototype there was some updates to height / min-height
for components that should spawn over the full height of the screen.

This introduced a regression that the sidebar and app content are
not scrollable.

Introduced in e67a63c2d6180c1485b72b7c5cd1ab4f821d7a64 which was
for sidebar only and followup of 02aba14057a1066aa8681ee5a001efd4c6fce5f1.

Other views or components that should be checked if scroll is
working as intended are:
- Bootloader view
- Dialog
- Guide

Other components seem to not have the scrolling issue as those have
either height or max-height set to 100vh and with that allow the
user to scroll the content.