backend/btc: add support for payment request transactions
Payment requests have been supported by bitbox02 firmware since v9.19.0.
This integrates the feature in the app, mocking a testnet payment request
in the Pocket widget section. Once the Pocket sell widget integration is
completed the mock function will be removed.

Note: to test this it is necessary to flash a dev device with a FW
having the test merchant enabled.