frontend: fix vertical alignment of radio buttons
Radio buttons and checkboxes vertical alignment was a bit off.

This change positions the form inputs to be on the baseline of the
label with normal font-size and lineheight.

With this some hacks can be removed and long text in labels are
more readable with normal lineheight.

If larger subheader font-size is used in the label, line-height 1
should be sufficient to vertically algin the input element.

Tested:
- UTXO selection in coin control
- script type in receive view
- backup checklist during create-wallet
- backup selection in restore-from-sdcard
- agree/understand checkbox in various views