frontend/exchange: refactor exchange deals logic
Exchange logic was largely implemented in the frontend. With this, all
the logic is moved to the backend, and the whole code is simplified.

Main changes:
- getExchangeDeals now has three params: coin, region and action
(buy/sell) and returns already filtered exchange deals, with best deal
flag already set
- mv sell.tsx buysell.tsx: this component now has an action param, that
allows to display different deals on the base of the enabled buy/sell
tab.
- all the /buy/ path has been moved to /exchange/
- a lot of frontend logic (e.g. to find the best deal) is not necessary
anymore and has been removed.
- pocket sell mockup button removed, as it is not necessary anymore

Still todo: we should check if the bb02 firmware supports silent
payments and can sell through Pocket using a backend endpoint instead
of just checking the version in the frontend. This will be done in a
future commit.