android: don't whitelist based on current app location
This is supposed to prevent external sites from loading within the
app, but we allowed all of them anyway when on the buy page to make
MoonPay work (we can't know the list of third party domains to
whitelist as it involves various third party credit card processors
and banks).

It's unlikely if this prevents any potential attacks, as frontend code
that is compromised can probably still call to external sites by
moving to the buy page and executing there.

The reason to remove this now is that webview.goBack() (history back)
does not work in combination with `loadDataWithBaseURL()` (it always
goes back to the first page). with `loadUrl()` history back works, but
onPageFinished() is not called for subsequent url changes.

Generally, this whitelisting has caused us a lot of pain and
engineering time, while being very fragile.