frontend: add and use ContentWrapper for Status outside of View
Using our new ContentWrapper in order for components that are
using Status and were outside of the content view,
to look like they are in the view.

It's done to improve aesthetics.

Also, they will be scrolling with the content instead of
"sticking on top" like it used to be.