frontend: created `GlobalBanners` and use it in several pages
created the new component `GlobalBanners`, extracting it out
from `app.tsx`. We use this component in `account.tsx`,
`accountsummary.tsx`, and all of the settings pages.

The reason for this is because we'd like to move them out of
`app.tsx` so they're in the document flow of the page instead of
living separately in `app.tsx` (visually placed above the page
view).