btc: refactor sighashes
The SigHashes struct is a precomputation of hashes to speed up the
computation of per-input sighashes. The idea is to compute it once,
not per-input.

We made the instance after creating the tx, before signing it. This
works fine, but with Silent Payments (BIP-352), the signer also
modifies the transaction: it generates the silent payment output.

THe pre-computed sighashes are not valid in that case, as the
transaction changes afterwards. We simply fix this by creating the
instance after signing in the BitBox02 (which will support silent
payments), and use it as before in the other keystores.