backend: update to version 4.44.0
Added missing CHANGELOG entry.
Reworded the Pocket sell CHANGELOG entry, the technical detail is
unimportant for the release note. Also the BTC are not sold *to* Pocket,
but Pocket sells it for the customer as a broker.

Lowercase bitcoin is normally used for the actual coins (uppercase
Bitcoin is used for the network/brand/...).