frontend: put send result in state
Put the result from the send request into the `Send` components state
and use it directly instead of `isSent` and `isAborted`. This simplifies
the component and will make the refactor to a functional component
easier because we'll be able to use a `useEffect` with a dependncy on
the `sendResult` state that can `switch` over it and also reset the
form.