frontend: remove signProgress from BB02 confirm
Right now the `BB02ConfirmSend` component subscribes to the
`signProgress` event. This event **can only be fired** if a BB01 is
connected: 8e307affe84f20e68f09a76d38eb636caa94724e it is not fired if a
BB02 is connected (searched the backend for `EventSignProgress` it is
only in the BB01 impl.).

Because the component is only shown if the connected device is
`bitbox02`, the state is always `undefined` and can be removed.